アップロード済の短縮URL定義ファイルを取得するコマンドを追加
